Refactor v2
