+++ package.json
@@ -75,6 +75,7 @@
         "@wharfkit/wallet-plugin-privatekey": "^1.0.0",
         "ajv": "^8.12.0",
         "atomicassets": "^1.5.1",
-        "idb-keyval": "^6.2.1"
+        "idb-keyval": "^6.2.1",
+        "zustand": "^4.5.2"
     }
 }
