+++ src/actions/atomic/getAccountAssets.ts
@@ -1,0 +1,15 @@
+import { Client } from "../../client";
+
+export const getAccountAssets = async (client: Client, account: string) => {
+  const { atomic } = client.network.config;
+  const { provider } = client;
+
+  const { rows } = await provider.v1.chain.get_table_rows({
+    code: atomic.atomicContract,
+    scope: account,
+    table: "assets",
+    limit: 100,
+  });
+
+  return rows;
+};
