+++ src/actions/atomic/getCollection.ts
@@ -1,0 +1,15 @@
+import { Client } from "../../client";
+
+export const getCollection = async (client: Client, collectionName: string) => {
+  const { atomic } = client.network.config;
+  const { provider } = client;
+
+  const { rows } = await provider.v1.chain.get_table_rows({
+    code: atomic.atomicContract,
+    scope: collectionName,
+    table: "collections",
+    limit: 100,
+  });
+
+  return rows;
+};
