+++ src/actions/index.ts
@@ -1,0 +1,34 @@
+export * from "./atomic/getAccountAssets";
+export * from "./atomic/getAsset";
+export * from "./atomic/getCollection";
+export * from "./atomic/getSchema";
+
+export * from "./dao/getAvatar";
+export * from "./dao/getDaoSettings";
+export * from "./dao/setAvatar";
+
+export * from "./ipfs/getIpfsResource";
+export * from "./ipfs/uploadIpfsResource";
+
+export * from "./tasks/getForceSettings";
+export * from "./tasks/getTask";
+export * from "./tasks/getRepetitions";
+export * from "./tasks/submitTask";
+export * from "./tasks/campaigns/createCampaign";
+export * from "./tasks/campaigns/getCampaigns";
+export * from "./tasks/batch/createBatch";
+export * from "./tasks/batch/getBatch";
+export * from "./tasks/reservations/getReservations";
+export * from "./tasks/reservations/reserveTask";
+
+export * from "./token/getDefiBoxPair";
+export * from "./token/swap";
+export * from "./token/getPrice";
+
+export * from "./vaccount/createAccount";
+export * from "./vaccount/deposit";
+export * from "./vaccount/getAccounts";
+export { getAvatar as getVAccountAvatar } from "./vaccount/getAvatar";
+export * from "./vaccount/getPendingPayout";
+export * from "./vaccount/payout";
+export * from "./vaccount/transfer";
