+++ src/actions/tasks/getRepetitions.ts
@@ -1,0 +1,20 @@
+import { Client } from "../../client";
+import { useEFXContracts } from "../../utils";
+
+export const getRepetitions = async (client: Client) => {
+  try {
+    const { tasks } = useEFXContracts(client);
+    const { provider } = client;
+
+    const response = await provider.v1.chain.get_table_rows({
+      code: tasks,
+      table: "repsdone",
+      scope: tasks,
+    });
+
+    return response.rows;
+  } catch (error) {
+    console.error(error);
+    throw error;
+  }
+};
