+++ src/index.ts
@@ -2,3 +2,4 @@ export * from "./constants";
 export * from "./client";
 export * from "./template";
 export * from "./types";
+export * from "./actions";
