--- src/services/dao.spec.ts
@@ -1,4 +1,0 @@
-import { describe, it } from "vitest";
-
-//TODO:: Add tests for DaoService
-describe("DaoService", async () => {});
