--- src/services/ipfs.spec.ts
@@ -1,4 +1,0 @@
-import { describe } from "vitest";
-
-//TODO:: Implement tests for IpfsService
-describe("IpfsService", async () => {});
