--- src/services/tasks.spec.ts
@@ -1,10 +1,0 @@
-import { describe, it } from "vitest";
-
-//TODO:: Implement tests for TasksService
-describe("TaskService", async () => {
-  it("should fetch task data correctly from IPFS", async () => {});
-  it("should fetch all campaigns", async () => {});
-  it("should fetch a specific campaign", async () => {});
-  it("should fetch a batch from a campaign", async () => {});
-  it("should throw an error if a user tries to reserve a task without logging in", async () => {});
-});
