--- src/services/token.spec.ts
@@ -1,4 +1,0 @@
-import { describe, it } from "vitest";
-
-//TODO:: Implement tests for TokenService
-describe("TokenService", async () => {});
