--- src/services/vaccount.spec.ts
@@ -1,4 +1,0 @@
-import { describe, it } from "vitest";
-
-//TODO:: Add tests for VAccountService
-describe("VAcccountService", async () => {});
